refactor : 불필요한 로그아웃 로직 수정 및 Hook으로 기능 분리
